requirements.txt
Django==2.2.3
python-slugify==3.0.2
pytz==2019.1
sqlparse==0.3.0
text-unidecode==1.2
